Allow the user to chose customs parameters before the template creation. The template owner must declare the parameters the user can chose in the templates.xml file. When the user wants to use such template, he has to fill the parameters. After that, the template is created with the user's inputs.
a

b

c

Allow the user to chose customs parameters before the template creation.
The template owner must declare the parameters the user can chose in
the templates.xml file. When the user wants to use such template, he
has to fill the parameters. After that, the template is created with
the user's inputs.

a

b

c